+++ src/main/java/com/raja/lib/invt/repository/PurchaseRepository.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Repository;
 import com.raja.lib.invt.model.Purchase;
 
 @Repository
-public interface PurchaseRepository extends JpaRepository<Purchase, Long> {
+public interface PurchaseRepository extends JpaRepository<Purchase, Integer> {
 	
 //	@Query("SELECT p FROM Purchase p JOIN FETCH p.ledger WHERE p.purchaseId = :purchaseId")
 //    Optional<Purchase> findByIdAndFetchLedgerEagerly(@Param("purchaseId") Long purchaseId);
